remove SignerGetters from setClientOptions
