+++ packages/widget/src/state/swapExecutionPage.ts
@@ -4,9 +4,11 @@ import { routeConfigAtom, skipRouteAtom } from "@/state/route";
 import { atom } from "jotai";
 import {
   DEEPLINK_CHOICE,
+  getConnectedSignersAtom,
   MinimalWallet,
   RECENT_WALLET_DATA,
   walletConnectDeepLinkByChainTypeAtom,
+  walletsAtom,
 } from "./wallets";
 import { atomEffect } from "jotai-effect";
 import { setTransactionHistoryAtom, transactionHistoryAtom } from "./history";
@@ -29,6 +31,11 @@ import {
 } from "@skip-go/client";
 import { currentPageAtom, Routes } from "./router";
 import { LOCAL_STORAGE_KEYS } from "./localStorageKeys";
+import { solanaWallets } from "@/constants/solana";
+import { getWallet, WalletType } from "graz";
+import { config } from "@/constants/wagmi";
+import { WalletClient } from "viem";
+import { getWalletClient } from "@wagmi/core";
 
 type ValidatingGasBalanceData = {
   chainId?: string;
@@ -364,6 +371,8 @@ export const skipSubmitSwapExecutionAtom = atomWithMutation((get) => {
   const getFallbackGasAmount = get(fallbackGasAmountFnAtom);
   const simulateTx = get(simulateTxAtom);
   const swapSettings = get(swapSettingsAtom);
+  const getSigners = get(getConnectedSignersAtom);
+  const wallets = get(walletsAtom);
 
   const { timeoutSeconds } = get(routeConfigAtom);
 
@@ -399,6 +408,43 @@ export const skipSubmitSwapExecutionAtom = atomWithMutation((get) => {
           simulate: simulateTx !== undefined ? simulateTx : route.sourceAssetChainId !== "984122",
           getFallbackGasAmount,
           ...submitSwapExecutionCallbacks,
+          getCosmosSigner: async (chainId) => {
+            if (getSigners?.getCosmosSigner) {
+              return getSigners.getCosmosSigner(chainId);
+            }
+            if (!wallets.cosmos) {
+              throw new Error("getCosmosSigner error: no cosmos wallet");
+            }
+            const wallet = getWallet(wallets.cosmos.walletName as WalletType);
+            if (!wallet) {
+              throw new Error("getCosmosSigner error: wallet not found");
+            }
+            const key = await wallet.getKey(chainId);
+
+            return key.isNanoLedger
+              ? wallet.getOfflineSignerOnlyAmino(chainId)
+              : wallet.getOfflineSigner(chainId);
+          },
+          getEvmSigner: async (chainId) => {
+            if (getSigners?.getEvmSigner) {
+              return getSigners.getEvmSigner(chainId);
+            }
+            const evmWalletClient = (await getWalletClient(config, {
+              chainId: parseInt(chainId),
+            })) as WalletClient;
+
+            return evmWalletClient;
+          },
+          getSvmSigner: async () => {
+            if (getSigners?.getSvmSigner) {
+              return getSigners.getSvmSigner();
+            }
+            const walletName = wallets.svm?.walletName;
+            if (!walletName) throw new Error("getSvmSigner error: no svm wallet");
+            const solanaWallet = solanaWallets.find((w) => w.name === walletName);
+            if (!solanaWallet) throw new Error("getSvmSigner error: wallet not found");
+            return solanaWallet;
+          },
         });
       } catch (error: unknown) {
         console.error(error);
